Update Choices and added School and Enrollment
This morning I added actual choice list tuples to use in conjuction with
charfields as the enum option may not be available to me but I want to
keep the choice option controls in the custom app solution.

I also finished adding the school_details model and the enrollment table.