Swapped over to a tabulator view using pandas dataframe.
We moved the filter system to its own function to stop repeating that code.
We also added pandas dataframe into the mix for handling the data.
We introduced tabulator for the table management because it gave some
sorting features without having to build them myself. Its not react quality
but it works for the point of the demo.

This is working but requires adding the feature for picking columns still...