Added an area for actions to the enhance_filter page.
I added an aside to the right with actions that can handle sending emails,
exporting to excel, exporting to csv, and printing mailing labels. The only
item we will be giving logic to right now is sending emails.

I also adjusted the page so that filters are 2/12 of the page, results is
8/12 and the actions is also 2/12 of the page.