Created email form for the demo.
I created a modal form for sending emails from the demo. It has options
for declaring to include parish, personal, or diocesan emails. It has the
subject input and body input as well. The To will be a service account for
the demo and recipients will be from the filtered options in results.

I will add the attachment system later.