Added Temp holding fields to models and finished import file.
Added temporary text fields for assignment history, additional languages,
and professional degrees to the Priest model.

This is to hold data that will be moved to a new model in the future.

Finished the import file to handle these new fields. And added logic to
handle all the nan values we were getting from the import file. Lastly,
I added the logic for handling the other assignments types.