Add 2 new feature checks.
We now have the system check if the user_obj.suspend == True, if yes it
denies the login and updates the log for user login attempts.

We also added an update for last_login on the user object so if sucessful
we log the last login on the account directly.