Added the feature suspend or unsuspend users.
In the Datagrid component we made it easier to mapping of columns.
We also accept options which allows for the suspend and other items editting
in the future from the table.

In the Auth service we created a handle for managing the suspend function.
We also introduce useMemo to allow better control of specific fields and
allows us to use OnChange and handlers in the table.